Time: 22 ms (99.38%), Space: 81.2 MB (33.99%) - LeetHub

+++ 2064-minimized-maximum-of-products-distributed-to-any-store/2064-minimized-maximum-of-products-distributed-to-any-store.java
@@ -17,7 +17,7 @@ class Solution {
     }
 
     public boolean isPossible(int[] quantities, int x, int n) {
-        int store = 0;
+        long store = 0;
         for (int quantity : quantities) {
             store += (quantity + x - 1) / x;
             if (store > n) {
